backend/requirements.txt
django>=5.0,<6
djangorestframework>=3.15
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.4.0
python-dotenv>=1.0
pytz
google-api-python-client>=2.147
google-auth>=2.33
google-auth-oauthlib>=1.2
gunicorn==22.0.0
requests==2.32.3
